refactor(numa): extract pegaflow-numa crate to decouple pegaflow-transfer from pegaflow-core
Move NumaNode, format_cpu_list, read_cpu_topology_from_sysfs and their
tests into a new pegaflow-numa crate. pegaflow-core re-exports these
symbols so downstream paths are unchanged. pegaflow-transfer now depends
only on pegaflow-numa instead of the full pegaflow-core.
