+++ pegaflow-numa/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "pegaflow-numa"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+log.workspace = true
