+++ pegaflow-transfer/Cargo.toml
@@ -8,7 +8,7 @@ parking_lot.workspace = true
 sideway.workspace = true
 rdma-mummy-sys = "0.2.3"
 log.workspace = true
-pegaflow-core.workspace = true
+pegaflow-numa.workspace = true
 logforth.workspace = true
 thiserror = "2.0"
 bytes = "1"
